Move the simple token parser into the String namespace
