+++ core-bundle/src/Resources/contao/modules/ModulePassword.php
@@ -11,7 +11,7 @@
 namespace Contao;
 
 use Contao\CoreBundle\OptIn\OptIn;
-use Contao\CoreBundle\Util\SimpleTokenParser;
+use Contao\CoreBundle\String\SimpleTokenParser;
 use Symfony\Component\HttpFoundation\Session\Session;
 
 /**
