+++ core-bundle/src/Resources/contao/modules/ModuleRegistration.php
@@ -12,7 +12,7 @@ namespace Contao;
 
 use Contao\CoreBundle\Exception\ResponseException;
 use Contao\CoreBundle\OptIn\OptIn;
-use Contao\CoreBundle\Util\SimpleTokenParser;
+use Contao\CoreBundle\String\SimpleTokenParser;
 
 /**
  * Front end module "registration".
